Updates along with the change in the type of item price

+++ src/main/java/jpabook/jpashop/controller/BookForm.java
@@ -1,6 +1,5 @@
 package jpabook.jpashop.controller;
 
-import jakarta.validation.constraints.Max;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -10,7 +9,7 @@ public class BookForm {
     private Long id;
 
     private String name;
-    private int price;
+    private float price;
     private int stockQuantity;
 
     private String author;
