+++ src/main/java/jpabook/jpashop/domain/item/Book.java
@@ -15,7 +15,7 @@ public class Book extends Item {
 
     private String isbn;
 
-    public static Book createBook(String name, int price, int stockQuantity, String author, String isbn) {
+    public static Book createBook(String name, float price, int stockQuantity, String author, String isbn) {
         Book book = new Book();
         book.setName(name);
         book.setPrice(price);
